Allow updating files on the boot partition
Until now OSTree copied only the vmlinuz and initramfs
binaries to the boot partition. This patch adds support
for copying other files from the /boot directory of the
updated tree.

How this works:

Ignore subdirectories, only files in root of the boot
directory are copied. There is overhead of copying files
to boot partition, therefore the amount of files in the
boot/ should be kept to the minimum and subdirectories
shouldn't really be necessary.

Files on the boot partition are updated only with major
releases, when kernel/initramfs bootcsum changes. Files
that require frequent updates should not be stored here.