Support for booting without initramfs
Previously when initramfs-* was not found in a deployment's
boot directory, it was assumed that rootfs is prepared for
ostree booting by a kernel patch.

With this patch, the behaviour changes to be - if initramfs-* not
found, assume that system is using a static ostree-prepare-root
as init process. Booting without initramfs is a common use case
on embedded systems. This approach is also more convenient than
having to patch the kernel.